requirements.txt
flask
pyyaml
pymongo
virtualenv
ansible